Move tests to own folder
