+++ .travis.yml
@@ -23,4 +23,5 @@ script:
   - flake8 *.py
   - flake8 config
   - flake8 cal
+  - flake8 tests
   - py.test
